fixed integer overflow for rle instructions

+++ src/pclint/Pcl.h
@@ -52,8 +52,8 @@ public:
   char keysep; // separate the instr signature from value/data so it may be interpreted as a string and used as key
   int32_t value;
   uint8_t * data;
-  uint8_t pos;
-  uint8_t limit;
+  uint16_t pos;
+  uint16_t limit;
   off64_t file_off;
 
   bool hasValue;
